requirements.txt
fastapi
uvicorn[standard]
databases[postgresql]
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
authlib
httpx
